requirements.txt
sklearn
numpy
pandas
pillow
newspaper3k
flask